refactor(hosted): colocate finalization state
